Illusi 14.0.0

+++ Illusive/src/illusi/src/sql/sql_backpack.ts
@@ -3,7 +3,7 @@ import { ExampleObj } from "../example_objs";
 import { Illusive } from "../../../illusive";
 import { SQLTrack, Track } from "../../../types";
 import { all_promises } from "../../../illusive_utilts";
-import { delete_track, insert_track, sql_track_to_track, track_from_uid, track_to_sqllite_insertion } from "./sql_tracks";
+import { delete_track, insert_track, sql_track_to_track, track_exists, track_from_uid, track_to_sqllite_insertion } from "./sql_tracks";
 import { sql_delete_from, sql_insert_values, sql_select, sql_where } from "./sql_utils";
 
 export async function empty_backpack(){
@@ -20,12 +20,13 @@ export async function toss_from_backpack(replacement_track: Track){
 }
 
 export async function backpack_tracks(){
-    const sql_tracks: SQLTrack[] = await db.getAllAsync(sql_select<Track>("tracks", "*"));
+    const sql_tracks: SQLTrack[] = await db.getAllAsync(sql_select<Track>("backpack", "*"));
     const tracks: Track[] = sql_tracks.map(sql_track => sql_track_to_track(sql_track));
     for(let i = 0; i < tracks.length; i++) tracks[i].playback!.artwork = Illusive.illusi_dark_icon;
     return tracks;
 }
 export async function add_to_backpack(uid: string){
+    if(await track_exists({uid, title:"", artists:[], duration:0})) return;
     const track = await track_from_uid(uid);
     await all_promises([
         delete_track(uid),
